fix up the forward query sending/handeling, fix some wrong freeing too,
Fri Jun 22 18:53:14 2001  George Lebl <jirka@5z.com>

	* daemon/xdmcp.c: fix up the forward query sending/handeling, fix
	  some wrong freeing too, and fix the hosts_ctl warning by including
	  a prototype.  We finally have chooser stuff all working! Yay!

	* daemon/xdmcp.c, gui/gdmchooser.c: we want to be very careful in
	  what types we pass xdmcp functions, the header does not include
	  areguments in the prototypes, so usage is snarfed from xdm

+++ gui/gdmchooser.c
@@ -279,7 +279,7 @@ gdm_chooser_xdmcp_discover (void)
     while (bl) {
 	ia = (struct in_addr *) bl->data;
 	sock.sin_addr.s_addr = ia->s_addr; 
-	XdmcpFlush (sockfd, &bcbuf, (XdmcpNetaddr) &sock, sizeof (struct sockaddr_in));
+	XdmcpFlush (sockfd, &bcbuf, (XdmcpNetaddr) &sock, (int)sizeof (struct sockaddr_in));
 	bl = bl->next;
     }
 
@@ -291,7 +291,7 @@ gdm_chooser_xdmcp_discover (void)
     while (ql != NULL) {
 	    ia = (struct in_addr *) ql->data;
 	    sock.sin_addr.s_addr = ia->s_addr;
-	    XdmcpFlush (sockfd, &querybuf, (XdmcpNetaddr) &sock, sizeof (struct sockaddr_in));
+	    XdmcpFlush (sockfd, &querybuf, (XdmcpNetaddr) &sock, (int)sizeof (struct sockaddr_in));
 	    ql = ql->next;
     }
